Add a Pull Request Template (#27)
## What Changed?

<!-- 
A summary of what you are proposing to change.
-->

Adds a template for GitHub pull requests as per [this
guide](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository).

## Why Does It Need To?

<!-- 
Describe the concern this change addresses. 

Reference any issues that are related and use closing words such as
"fixes" or
"closes" if those issues would be fully addressed by this PR. 

You can find out more about closing words here:

https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/using-keywords-in-issues-and-pull-requests

Try to make your PR about one single concern or issue, unless they are
mutually dependent.
-->

Since we now require PRs every time we merge to main it makes sense to
provide a bit of guidance on what they should look like. Some of the
basic ideas here are adapted from Kani's PR template.

In fact this PR uses the template it is trying to merge, so have a look.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [ ] Documentation in Notion has been updated
- [x] Tests for new behaviors are provided
  - [x] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.