A much simpler way to set the sysroot. (#34)
## What Changed?

Reimplements #32 in a simpler manner. Partially reverts those changes
but with the same overall outcome.

Also added documentation for what is happening.

## Why Does It Need To?

The other version to achieve this needed more changes to `rustc_plugin`
that were haphazard and would have been annoying to try and get merged
eventually. This way is slightly more hacky but with a *significantly*
reduced footprint. Everything is now contained in `dfpp` alone.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [x] Documentation in Notion has been updated
- [x] Tests for new behaviors are provided
  - [x] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.