Fix refs on Identifier (#36)
## What Changed?

It makes the interface of `Context` consistently use `Identifier`, not
`&Identifier`

## Why Does It Need To?

The prior API is odd and inconsistent. And since `Identifier` implements
`Copy` we don't need to pass it by reference. This should make it easier
to use the API, because you don't need to carry around lifetimes for
identifiers.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [x] Documentation in Notion has been updated
- [x] Tests for new behaviors are provided
  - [x] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.