Remove Laziness for Markers (#89)
## What Changed?

Eagerly visit all IDs in the program under analysis and register
potential markers.
Rips out all the infrastructure that was used to do it lazily.

Refactors the `CollectingVisitor` into two structs, the visitor and then
an `SPDGGenerator` that runs the actual analysis. This separates the
concerns out better. The former stores (among other things) a mutable
`MarkerDatabase` that it fills with the markers for local functions and
types and it collects a vector of analysis targets.
Once the visit is done it is converted into the `SPDGGenerator` which
stores the read-only `MarkerCtx`.

## Why Does It Need To?

Lazy markers were confusing. Markers on controllers were never
registered. Also could cause surprising vacuity if markers were never
encountered because e.g. their call sites happened in a closure.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [x] Documentation in Notion has been updated
  - [x] Remove from list of arguments
  - [x] Remove from step-by-step guide
- [x] Tests for new behaviors are provided
  - [x] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.