Introducing the Paralegal proc-macro library (#96)
## What Changed?

Introduces the implicit `cfg(paralegal)` option enabled when running
`paralegal_flow` and creates macro versions of the paralegal annotations
that can be used in any code, regardless of whether `register_tool` is
used.

Also migrates test cases using `label` to the `marker` annotation.

## Why Does It Need To?

This is a further step in making Paralegal compatible with projects that
use other (and critically non-nightly) toolchains.

So far you can run Paralegal on code bases that do not use a nightly
compiler. However to add annotations you need to enable the
`register_tool` feature, which means that the code is no longer
compilable with a stable compiler.

With this change you can instead link against the `paralegal` library
which encapsulates all the attribute shenanigans.

`paralegal-flow` now implicitly passes `--cfg paralegal` to the compiler
as well as the `feature(register_tool)` and
`register_tool(paralegal_flow)`. This allows users to conditionally
include code only when paralegal analysis is enabled (with
`cfg!(paralegal)`).

The library uses this mechanism to toggle the implementations for its
attributes. In any non-paralegal setting the attributes are no-ops.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [ ] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [ ] Documentation in Notion has been updated
- [ ] Tests for new behaviors are provided
  - [ ] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.