Add test normalization to prevent hard errors in function resolution (#107)
## What Changed?

Add a test normalization for generic arguments when resolving call
instances.

## Why Does It Need To?

Instance resolution can fail with a hard error if the generic arguments
can't be resolved. We don't want that to be a hard error since it can
arise from legitimate situations during our analysis. So we try the
resolution separately before and catch it, reporting a graceful warning
and carrying on.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [ ] Documentation in Notion has been updated
- [ ] Tests for new behaviors are provided
  - [ ] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.