Update the marker assignment policy (#139)
## What Changed?

Updates the marker assignment algorithm. It now attaches to a node `n`

1. all markers for `typeof(n)`
2. If `children(n).is_empty()` (determined by flowistry’s `PlaceInfo`)
then we must be a terminal place where subplaces aren’t tracked, meaning
that they won’t be marked via rule 1 and we have to overtaint, so
include all markers from `subtypes(typeof(n))`
3. Add all markers from `n.iter_projections().map(typeof)` to mark
ourselves if we are a field of a tainted type.

## Why Does It Need To?

This implements an updated marker policy. It simplifies the semantics
and makes them sound hopefully. This is possible now because we have
much better field sensitivity. Markers are now attached to every node of
the marked type as well as
- nodes which are children (e.g. fields) of a marked type
- nodes which are parents of the marked type **if** it the children of
this node are not separately tracked (overapproximation)

Mostly implements #129 

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [ ] Documentation in Notion has been updated
- [x] Tests for new behaviors are provided
  - [ ] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.

---------

Co-authored-by: Will Crichton <wcrichto@cs.stanford.edu>