Merge the pdg construction from flowistry into this repo (#140)
## What Changed?

Merges the flowistry PDG into this repo. Contains updates to the
handling of async closures to have better field sensitivity.

Caveats: This merges an alias analysis bug #144. But I need main to
catch up and this bug does not appear to influence any use cases so I am
going ahead with this merge.

## Why Does It Need To?

This allows us to iterate faster on the actual construction algorithm.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [ ] Documentation in Notion has been updated
- [ ] Tests for new behaviors are provided
  - [ ] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.