Optimize graph construction (#150)
## What Changed?

Splits the PDG construction into two passes. 

1. A local (per procedure), fixpoint analysis that determines the "last
mutation" of every place at every point in the program.
2. A global pass that constructs a PDG by combining the local results.

## Why Does It Need To?

Improves performance by shrinking the state carried during the fixpoint
analysis. This is achieved in two ways:
1. The state now only tracks the last mutation but not edge and node
sets.
2. The state only applies to the local procedure.

Furthermore this is a step towards modularizing the PDG construction (to
a per-procedure model). This will enable caching and cross crate
analysis.

## Caveats

Removes the `FakeEffect` API. It is currently unused and it is not clear
that it will be required after the truly modular analysis has been
implemented.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [ ] Documentation in Notion has been updated
- [ ] Tests for new behaviors are provided
  - [ ] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.

---------

Co-authored-by: Carolyn Zech <carolynzech@gmail.com>
Co-authored-by: Will Crichton <wcrichto@cs.stanford.edu>