Support for CodeLLDB debugger (#156)
## What Changed?

Adds the option to attach remotely to a CodeLLDB debugging session
directly from within paralegal-flow. It uses the mechanism described
[here](https://github.com/vadimcn/codelldb/blob/v1.10.0/MANUAL.md)

Adds a `--attach-to-debugger` command line option which accepts the name
of a debugger to initiate a session with. Currently only the value
`code-lldb` is accepted. It will attach to a debugger for every analysis
target. Skipped crates (where regular rustc runs) do not attach.

## Why Does It Need To?

Vastly simplifies the process of using a debugger with the tool.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [ ] Documentation in Notion has been updated
- [ ] Tests for new behaviors are provided
  - [ ] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.