Implement serialization protection via magic (#169)
## What Changed?

This adds a u64 magic number to the beginning of the output artifact
from `paralegal-flow` which is verified upon loading the artifact.

The magic number is a hash of the modification dates of all files in the
`paralegal-spdg` crate.

## Why Does It Need To?

At present there is no protection that the format expected by a policy
application conforms to the format written by the flow analyzer.
Specifically the analyzer may be compiled against an older or newer
version of `paralegal-spdg`. In such cases the derived `serde`
implementation for the output artifact may differ. `bincode`, the
serialization library we use, does not offer protection against this
case, leading to out-of-memory errors and memory leaks in
deserialization.

The magic hash added detects this version mismatch and reports it.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [ ] Documentation in Notion has been updated
- [ ] Tests for new behaviors are provided
  - [ ] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.