Automated Marking of Side Effects and Data Leakage (#184)
## What Changed?

Adds optional automatic marking for potential side effects. The idea is
based on the [scrutinizer tool](https://github.com/artemagvanian/pear).

The following markers get attached automatically if the
`--side-effect-markers` option is enabled.

- `auto:side-effect:raw-ptr`: A raw pointer dereference was encountered
and can cause data leakage
- `auto:side-effect:transmute`: A coercion was encountered that could
involve pointers which can later be used to potentially leak data
- `auto:side-effect:intrinsic`: An unsafe intrinsic was encountered that
could cause data leakage (e.g. memory copy)
- `auto:side-effect:unknown:virtual`: A virtual function could not be
resolved and is conservatively assumed to perform side effects.
- `auto:side-effect:foreign`: An `extern` function is encountered and
must conservatively be assumed to perform side effects
- `auto:side-effect:unknown:fn-ptr`: A function pointer was encountered
and must conservatively be assumed to perform side effects
- `auto:side-effect:unknown`: Fallback marker for unsupported code

This is supposed to be used in conjunction with the markers from
`stdlib-markers.toml` which form an allowlist for stdlib functions that
use low-level code which would be conservatively assumed to be side
effects but we allow them for stdlib functions (such as `Vec`) that are
known to be safe.

This also includes a new internal feature
(`_internal_on_all_module_children`) for external marker files where a
marker can be applied to an entire module or the `impl`s of a struct.

Includes most of the tests from the original scrutinizer.

## Why Does It Need To?

Improves the soundness by letting policies reason about all potential
side effects, not just those of explicitly marked functions.

## Checklist

- [x] Above description has been filled out so that upon quash merge we
have a
  good record of what changed.
- [x] New functions, methods, types are documented. Old documentation is
updated
  if necessary
- [ ] Documentation in Notion has been updated
- [x] Tests for new behaviors are provided
  - [ ] New test suites (if any) ave been added to the CI tests (in
`.github/workflows/rust.yml`) either as compiler test or integration
test.
*Or* justification for their omission from CI has been provided in this
PR
    description.

---------

Co-authored-by: Justus Adam <justus@macbook-air.mynetworksettings.com>